+++ notary/src/main/kotlin/com/ing/zknotary/gradle/zinc/template/parameters/LinearPointerTemplateParameters.kt
@@ -5,7 +5,7 @@ import com.ing.zknotary.gradle.zinc.template.TemplateParameters
 
 object LinearPointerTemplateParameters : TemplateParameters(
     "linear_pointer.zn",
-    listOf(StringTemplateParameters(LinearPointerSurrogate.MAX_CLASS_NAME_SIZE.toShort()))
+    listOf(StringTemplateParameters(LinearPointerSurrogate.MAX_CLASS_NAME_SIZE))
 ) {
     override fun getReplacements() = mapOf(
         "CLASS_NAME_STRING_SIZE" to LinearPointerSurrogate.MAX_CLASS_NAME_SIZE.toString(),
