+++ notary/src/main/kotlin/com/ing/zknotary/gradle/zinc/template/parameters/X500PrincipalTemplateParameters.kt
@@ -5,7 +5,7 @@ import com.ing.zknotary.gradle.zinc.template.TemplateParameters
 
 object X500PrincipalTemplateParameters : TemplateParameters(
     "x500_principal.zn",
-    listOf(StringTemplateParameters(X500PrincipalSurrogate.PRINCIPAL_SIZE.toShort()))
+    listOf(StringTemplateParameters(X500PrincipalSurrogate.PRINCIPAL_SIZE))
 ) {
     override fun getReplacements() = mapOf(
         "PRINCIPAL_NAME_STRING_SIZE" to X500PrincipalSurrogate.PRINCIPAL_SIZE.toString(),
